Struct treexml::Element

pub struct Element {
    pub prefix: Option<String>,
    pub name: String,
    pub attributes: HashMap<String, String>,
    pub children: Vec<Element>,
    pub contents: Option<String>,
}

An XML element

Fields

prefix

Tag prefix, used for namespacing: xsl in xsl:for-each

name

Tag name: for-each in xsl:for-each

attributes

Tag attributes

children

A vector of child elements

contents

Contents of the element

Methods

impl Element

fn new<S>(name: S) -> Element where S: Into<String>

Create a new Element with the tag name name

fn find_child<P>(&self, predicate: P) -> Option<&Element> where P: for<'r> Fn(&'r &Element) -> bool

Find a single child of the current Element, given a predicate

fn find_child_mut<P>(&mut self, predicate: P) -> Option<&mut Element> where P: for<'r> FnMut(&'r &mut Element) -> bool

Find a single child of the current Element, given a predicate; returns a mutable borrow

fn filter_children<P>(&self, predicate: P) -> Filter<Iter<Element>, P> where P: for<'r> Fn(&'r &Element) -> bool

Filters the children of the current Element, given a predicate

fn filter_children_mut<P>(&mut self, predicate: P) -> Filter<IterMut<Element>, P> where P: for<'r> FnMut(&'r &mut Element) -> bool

Filters the children of the current Element, given a predicate; returns a mutable iterator